feat(modularization): Phase 3 REST execute in clawql-api, vault primitives in memory
Move rest-operation into clawql-api and call it directly from execute-core,
shrinking ExecuteEnvironment to native GraphQL/gRPC and in-process GraphQL only.
Extract vault config/utils/slug-index and slugifyTitle into clawql-memory with
subpath exports so MCP startup avoids loading the full memory barrel.
