+++ packages/clawql-api/src/execute/index.ts
@@ -1,4 +1,5 @@
 export * from "./execute-core.js";
 export * from "./execute-live.js";
 export * from "./field-projection.js";
+export * from "./rest-operation.js";
 export * from "./types.js";
