+++ packages/clawql-memory/src/index.ts
@@ -1,4 +1,8 @@
 export * from "./chunk/memory-chunk.js";
 export * from "./cache/artifact-cache.js";
 export * from "./ingest-file.js";
+export * from "./ingest/slug.js";
+export * from "./vault/config.js";
 export * from "./vault/markdown.js";
+export * from "./vault/slug-index.js";
+export * from "./vault/utils.js";
