+++ src/vault-config.ts
@@ -1,14 +1,12 @@
 /**
- * Obsidian vault path from CLAWQL_OBSIDIAN_VAULT_PATH (optional).
- * When set, the MCP server validates the path at startup.
+ * Obsidian vault startup validation (MCP transport). Path resolution lives in `clawql-memory/vault/config`.
  */
 
 import { constants } from "node:fs";
 import { access, stat } from "node:fs/promises";
-import { resolve } from "node:path";
+import { DEFAULT_OBSIDIAN_VAULT_PATH, getObsidianVaultPath } from "clawql-memory/vault/config";
 
-/** Default mount path in Docker / NAS deployments. */
-export const DEFAULT_OBSIDIAN_VAULT_PATH = "/vault";
+export { DEFAULT_OBSIDIAN_VAULT_PATH, getObsidianVaultPath };
 
 export type VaultStartupStatus = {
   configured: boolean;
@@ -25,18 +23,6 @@ let vaultStartupStatus: VaultStartupStatus = {
   writable: true,
   degraded: false,
 };
-
-/**
- * Resolved Obsidian vault directory, or `null` if vault integration is disabled
- * (`CLAWQL_OBSIDIAN_VAULT_PATH` unset or empty).
- */
-export function getObsidianVaultPath(): string | null {
-  const raw = process.env.CLAWQL_OBSIDIAN_VAULT_PATH?.trim();
-  if (raw === undefined || raw === "") {
-    return null;
-  }
-  return resolve(raw);
-}
 
 /**
  * Ensures the vault path exists, is a directory, and is readable + writable.
